sinaPatent/requirements.txt


# 测试工具
coverage>=7.3.0
pytest>=7.4.0
pytest-django>=4.7.0

# 代码质量
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# 性能优化


# 安全


# 环境变量
python-decouple>=3.8

# 日志
structlog>=23.1.0

# 时间处理
python-dateutil>=2.8.2

# HTTP客户端
requests>=2.31.0

# 数据验证
marshmallow>=3.20.0

# 缓存
redis>=5.0.0

# 数据库
psycopg2-binary>=2.9.7

# 静态文件
whitenoise>=6.6.0

# 生产环境
gunicorn>=21.2.0

# 文本处理和TF-IDF
scikit-learn>=1.3.0
jieba>=0.42.1
numpy>=1.24.0
pandas>=2.0.0

# 网页解析
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 网络请求
urllib3>=2.0.0
 
